QSP-7 Missing check if SimpleBuyout is vault owner
Added a function SimpleVault.vaultOwner(), and checked SimpleBuyout's ownership of the Vault in the SimpleBuyout.placeBid() function. The reason for adding an extra function instead of using the existing Vault.owner() was that IVault interface needed a non-overridable owner() equivalent, as the Ownable contract from OpenZeppelin does not specify owner() function as virtual, unlike the transferOwnership() function.